# Backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "ระบบจัดการสต็อกสินค้า",
        "title": "Inventory API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/api/admin/products/import": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Import products using CSV file (Admin only)\nCSV format example:\n\nname,price,stock\nKeyboard,20000,10\nMouse,10000,5\nMonitor,150000,3",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Products Import/Export"
                ],
                "summary": "Import products (admin)",
                "parameters": [
                    {
                        "type": "file",
                        "description": "CSV file to import",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
        "/api/products": {
            "get": {
                "description": "ดึงรายการสินค้า ถ้าเป็น Admin จะเห็นของทุกคน ถ้าเป็น User จะเห็นแค่ของตัวเอง",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Products"
                ],
                "summary": "ดูสินค้า (Admin เห็นทั้งหมด / User เห็นแค่ของตัวเอง)",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Product"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "สร้างสินค้าโดยส่ง JSON เข้ามา (ระบบจะผูกกับ User ที่ Login อัตโนมัติ)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Products"
                ],
                "summary": "เพิ่มสินค้าใหม่",
                "parameters": [
                    {
                        "description": "ข้อมูลสินค้า",
                        "name": "product",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Product"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Product"
                        }
                    }
                }
            }
        },
...
        "/api/products/import": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "multipart/form-data"
                ],
                "tags": [
                    "Products Import/Export"
                ],
                "summary": "Import products (User)",
                "parameters": [
                    {
                        "type": "file",
                        "description": "CSV file",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/api/products/import/preview": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Preview product data from CSV / Excel file before importing (no database insert)\n\nSupported file types:\n- CSV (.csv)\n- Excel (.xlsx, .xls)\n\nExpected columns:\nname, code, price, stock",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Products Import/Export"
                ],
                "summary": "Preview import products",
                "parameters": [
                    {
                        "type": "file",
                        "description": "CSV / Excel file to preview",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Preview result with total and preview rows",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/products/import/template": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Download Excel template for importing products\n\nColumns:\nname | price | stock\n\nNotes:\n- code can be empty (auto-generated by system)\n- ID is auto-generated by database",
                "produces": [
                    "application/vnd.openxmlformats-officedocument.spreadsheetml.sheet"
                ],
                "tags": [
                    "Products Import/Export"
                ],
                "summary": "Download product import template",
                "responses": {
                    "200": {
                        "description": "Excel template file",
                        "schema": {
                            "type": "file"
                        }
                    }
                }
            }
        },
        "/api/products/{id}": {
            "put": {
                "description": "Update product details by ID (ต้องเป็นเจ้าของเท่านั้น)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Products"
                ],
                "summary": "แก้ไขสินค้า",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Product ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Product Data",
                        "name": "product",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Product"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Product"
                        }
                    }
                }
            },
            "delete": {
                "description": "ลบสินค้าตาม ID (เฉพาะ Admin เท่านั้นที่ลบได้ ลบได้ทุกชิ้นไม่ว่าของใคร)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Products"
                ],
                "summary": "ลบสินค้า (Admin Only)",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Product ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "models.Product": {
            "type": "object"
        },
...
    }
}