# Backend/handlers/product_import.go
package handlers

import (
	"strconv"

	"github.com/gofiber/fiber/v2"
	"github.com/inventory-backend/database"
	"github.com/inventory-backend/models"
	"github.com/inventory-backend/utils"

	"github.com/xuri/excelize/v2"
)

// ImportProducts godoc
// @Summary      Import products (admin)
// @Description  Import products using CSV file (Admin only)
//
// @Description  CSV format example:
// @Description
// @Description  name,price,stock
// @Description  Keyboard,20000,10
// @Description  Mouse,10000,5
// @Description  Monitor,150000,3
//
// @Tags         Products Import/Export
// @Accept       multipart/form-data
// @Produce      json
// @Param        file  formData  file  true  "CSV file to import"
// @Success      200   {object}  map[string]interface{}
// @Failure      400   {object}  map[string]string
// @Security     BearerAuth
// @Router       /api/admin/products/import [post]
func ImportProducts(c *fiber.Ctx) error {
	userID := getUserID(c)

	rows, err := readImportFile(c)
	if err != nil {
		return err
	}

	success := 0
	failed := 0

	for i, row := range rows {
		if i == 0 {
			continue // skip header
		}

		if len(row) < 3 {
			failed++
			continue
		}

		price, err1 := strconv.ParseFloat(row[1], 64)
		stock, err2 := strconv.Atoi(row[2])
		if err1 != nil || err2 != nil {
			failed++
			continue
		}

		product := models.Product{
			Name:   row[0],
			Price:  price,
			Stock:  stock,
			Code:   utils.GenerateProductCode(),
			UserID: userID,
		}

		if err := database.DB.Create(&product).Error; err != nil {
			failed++
		} else {
			success++
		}
	}

	return c.JSON(fiber.Map{
		"message": "Import completed",
		"success": success,
		"failed":  failed,
	})
}

// @Summary Import products (User)
// @Tags Products Import/Export
// @Security BearerAuth
// @Accept multipart/form-data
// @Param file formData file true "CSV file"
// @Success 200 {object} map[string]interface{}
// @Router /api/products/import [post]
func ImportProductsForUser(c *fiber.Ctx) error {
	userID := getUserID(c)
	if userID == 0 {
		return c.Status(401).JSON(fiber.Map{"error": "Unauthorized"})
	}

	rows, err := readImportFile(c)
	if err != nil {
		return err
	}

	success := 0
	failed := 0

	for i, row := range rows {
		if i == 0 {
			continue
		}

		if len(row) < 3 {
			failed++
			continue
		}

		price, err1 := strconv.ParseFloat(row[1], 64)
		stock, err2 := strconv.Atoi(row[2])
		if err1 != nil || err2 != nil {
			failed++
			continue
		}

		product := models.Product{
			Name:   row[0],
			Price:  price,
			Stock:  stock,
			Code:   utils.GenerateProductCode(),
			UserID: userID,
		}

		if err := database.DB.Create(&product).Error; err != nil {
			failed++
		} else {
			success++
		}
	}

	return c.JSON(fiber.Map{
		"message": "Import completed",
		"success": success,
		"failed":  failed,
	})
}

// PreviewImportProducts godoc
// @Summary      Preview import products
// @Description  Preview product data from CSV / Excel file before importing (no database insert)
// @Description
// @Description  Supported file types:
// @Description  - CSV (.csv)
// @Description  - Excel (.xlsx, .xls)
// @Description
// @Description  Expected columns:
// @Description  name, price, stock (code is optional and auto-generated)
// @Tags         Products Import/Export
// @Accept       multipart/form-data
// @Produce      json
// @Param        file  formData  file  true  "CSV / Excel file to preview"
// @Success      200   {object}  map[string]interface{} "Preview result with total and preview rows"
// @Failure      400   {object}  map[string]string
// @Security     BearerAuth
// @Router       /api/products/import/preview [post]
func PreviewImportProducts(c *fiber.Ctx) error {
    rows, err := readImportFile(c)
    if err != nil {
        return c.Status(400).JSON(fiber.Map{"error": err.Error()})
    }

    preview := []fiber.Map{}

    for i, row := range rows {
        if i == 0 {
            continue // skip header
        }
        
        if len(row) < 3 {
            continue // skip ถ้าไม่มีเพียงพอ
        }

        preview = append(preview, fiber.Map{
            "name":  row[0],
            "price": row[1],
            "stock": row[2],
        })
    }

    return c.JSON(fiber.Map{
        "total":   len(preview),
        "preview": preview,
    })
}

// DownloadProductTemplate godoc
// @Summary      Download product import template
// @Description  Download Excel template for importing products
// @Description
// @Description  Columns:
// @Description  name | price | stock
// @Description
// @Description  Notes:
// @Description  - code can be empty (auto-generated by system)
// @Description  - ID is auto-generated by database
// @Tags         Products Import/Export
// @Produce      application/vnd.openxmlformats-officedocument.spreadsheetml.sheet
// @Success      200  {file}  file "Excel template file"
// @Security     BearerAuth
// @Router       /api/products/import/template [get]
func DownloadProductTemplate(c *fiber.Ctx) error {
    f := excelize.NewFile()
    sheet := "Template"
    f.SetSheetName("Sheet1", sheet)

    headers := []string{"name","price", "stock"}
    for i, h := range headers {
        cell, _ := excelize.CoordinatesToCellName(i+1, 1)
        f.SetCellValue(sheet, cell, h)
    }

    f.SetCellValue(sheet, "A2", "Keyboard")
    f.SetCellValue(sheet, "B2", 20000)
    f.SetCellValue(sheet, "C2", 10)

    c.Set("Content-Type", "application/vnd.openxmlformats-officedocument.spreadsheetml.sheet")
    c.Set("Content-Disposition", "attachment; filename=products_template.xlsx")

    return f.Write(c.Response().BodyWriter())
}